Break out the reinstate and offer new course to separate controllers
This also updates the path helpers to use the correct convention.
