+++ app/components/support_interface/application_choice_component.html.erb
@@ -5,9 +5,9 @@
         <ul class="app-summary-card__actions-list">
           <div class="app-summary-card__actions">
             <%= govuk_link_to 'Change offered course',
-              support_interface_application_form_change_offered_course_search_path(
+              support_interface_application_form_application_choice_change_offered_course_search_path(
                 application_form_id: application_choice.application_form.id,
-                application_choice_id: application_choice.id
+                application_choice_id: application_choice.id,
               ) %>
           </div>
         <ul>
