+++ config/routes.rb
@@ -810,17 +810,17 @@ Rails.application.routes.draw do
       get '/right-to-work-or-study' => 'application_forms/right_to_work_or_study#edit', as: :application_form_edit_right_to_work_or_study
       patch '/right-to-work-or-study' => 'application_forms/right_to_work_or_study#update'
 
-      get '/reinstate-offer/:application_choice_id' => 'application_forms/application_choices#confirm_reinstate_offer', as: :application_form_reinstate_offer
-      patch '/reinstate-offer/:application_choice_id' => 'application_forms/application_choices#reinstate_offer'
+      get '/reinstate-offer/:application_choice_id' => 'application_forms/application_choices/reinstate_declined_offer#confirm_reinstate_offer', as: :application_form_application_choice_reinstate_offer
+      patch '/reinstate-offer/:application_choice_id' => 'application_forms/application_choices/reinstate_declined_offer#reinstate_offer'
 
-      get '/change-offered-course-search/:application_choice_id' => 'application_forms/application_choices#change_offered_course_search', as: :application_form_change_offered_course_search
-      post '/change-offered-course-search/:application_choice_id' => 'application_forms/application_choices#search'
+      get '/change-offered-course-search/:application_choice_id' => 'application_forms/application_choices/change_offered_course#change_offered_course_search', as: :application_form_application_choice_change_offered_course_search
+      post '/change-offered-course-search/:application_choice_id' => 'application_forms/application_choices/change_offered_course#search'
 
-      get '/choose-offered-course/:application_choice_id' => 'application_forms/application_choices#offered_course_options', as: :application_form_choose_offered_course_option
-      post '/choose-offered-course/:application_choice_id' => 'application_forms/application_choices#choose_offered_course_option'
+      get '/choose-offered-course/:application_choice_id' => 'application_forms/application_choices/change_offered_course#offered_course_options', as: :application_form_application_choice_choose_offered_course_option
+      post '/choose-offered-course/:application_choice_id' => 'application_forms/application_choices/change_offered_course#choose_offered_course_option'
 
-      get '/confirm-offered-course/:application_choice_id/:application_choice_id' => 'application_forms/application_choices#confirm_offered_course_option', as: :application_form_confirm_offered_course_option
-      patch '/confirm-offered-course/:application_choice_id/:application_choice_id' => 'application_forms/application_choices#update_offered_course_option'
+      get '/confirm-offered-course/:application_choice_id/:application_choice_id' => 'application_forms/application_choices/change_offered_course#confirm_offered_course_option', as: :application_form_application_choice_confirm_offered_course_option
+      patch '/confirm-offered-course/:application_choice_id/:application_choice_id' => 'application_forms/application_choices/change_offered_course#update_offered_course_option'
     end
 
     get '/ucas-matches' => 'ucas_matches#index'
